core/browser_source_core_module/requirements.txt
# WaddleBot Flask Module - Dependencies

# Core Framework
quart>=0.20.0
hypercorn>=0.16.0

# HTTP Client (async)
httpx>=0.27.0,<0.28.0

# gRPC
grpcio>=1.67.0
grpcio-tools>=1.67.0

# JWT
PyJWT>=2.8.0

# Environment
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
